requirements.txt
certifi
cffi
chardet
greenlet
idna
mysql-connector-python
numpy
opencv-python
packaging
Pillow
protobuf
pycparser
pyparsing
PyQt5
PyQt5-Qt5
PyQt5-sip
python-dotenv
pytz
requests
six
sounddevice
soupsieve
SQLAlchemy
tzlocal
urllib3
